+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-__version__ = "1.1.1"
+__version__ = "1.2.0"