+++ src/titiler/core/titiler/core/__init__.py
@@ -1,6 +1,6 @@
 """titiler.core"""
 
-__version__ = "1.1.1"
+__version__ = "1.2.0"
 
 from . import dependencies, errors, factory, routing  # noqa
 from .factory import (  # noqa